fix curl transport command
Currently the  content-length field is missing in the request header. To fix this we have to use an other parameter set for the curl PUT request. 
The difference between '--file-upload -' and '-X PUT --data-binary @-' is that first argument set  streams directly to the request without buffering.  Therefore the content-length  header field is left out in the  request . The later one buffers the input.