ad_mu - Disable removing stale lock
LockFile::Simple in default configuration removes stale locks after 3600
seconds, even though "stale" option is set to false. To not remove any
stale lock, option "hold" have to be set to zero as well.

For ad_mu service the lock is used to indicate error in previous run,
which have to be resolved manually before next run of the service.
Therefore any automatic removal of stale lock is unacceptable.