Generic_send supports propagation to Windows
Two destination types were added to generic_send: user@host-windows and
host-windows-proxy.
user@host-windows will be used as a direct connection to windows server
2019 through ssh. Tar file, which will be sent, is encoded to base64 in
generic_send.
host-windows-proxy will be used as a connection to windows proxy server
(solution for older servers). Tar file, which will be sent, is also encoded to base64 in
generic_send, but in its end is added destination server. Name of the
proxy server has to be defined in the file
/etc/perun/services/generic_send/generic_send.conf in the variable
WINDOWS_PROXY.