refactor(send): removed lock from ad_mu service
Service ad_mu used error lock to prevent starting after the previous run
failed. It was used because the service worked with cache files which
could contain wrong data, so the next run could compute incorect results.
However, cached files are not used anymore, so the lock is
pointless. Therefore, it was removed from the script.