feat(o365_mu): removing and preserving resource mailboxes
- Because of technical issues with updating resource mailboxes data, we
were unable to finish migration of last version of this service to
production envrionment. The logic about resource mailboxes should be
removed, but there is a plan to further work on it in the future. For
this reason it was preserved and only removed from main code.
- Main parts of the code were commented and they are not processed any
more.
- All methodes and variables were preserved (but we are not using them)
- Sorting of emails were removed, because of cache in send script.
Cached data are not sorted so the record from gen script and the record
from cache would differ even if they are same in the context of content.