refactor(o365_mu): Processing bulk of entities
 - gen script is still the same; only the protocol has been increased (but
it has no direct impact on the service or the propagation process)
 - send script connector was merged with the process script to the one
working script (there is no need for multi-thread application anymore)
- new o365_mu_process script is now completing all logic of the client
site of the service. It is sending several bulks (every bulk has
a predefined number of entities inside) and processing every bulk on the
server site separately. When bulk is processed on the sever side, the
client will resolve it and save new state of objects to the cache.
The script would end with an error if any object was processed
unsuccessfully.